finished up Clearings, started Player Classes
Added the last major useful methods to the Clearing Class.
Started working on the Player class; I'm thinking of adding each faction as a subclass of the Clearing, each with its own methods to manipulate its 'faction board'. The actual logic of the game will still have to be added on the outside of these classes.