get wood available method and others
Added a few smaller methods to the Clearing / Board classes and made decks shuffle themselves after adding cards to them.

The big change was implementing the get_wood_available method, which uses a graph algorithm to find all of the groups of clearings ruled by the Marquise to find out how much wood is available to build with in each clearing.