resolving actions vs finding next action
This is work completed a few days ago. This is easily the most complicated section I have worked on thus far for this project.
Starting here, I am separating the step function into two main ideas:
- One function that, when given an action, performs EXACTLY what needs to be done for that one action, with little to nothing more
- A second function that, given the current state of the game, runs the game forward until the next choice is possible. We then break out and wait for that player to make their choice.

Essentially, the exact state of the game / the current player's turn is exactly represented by the Board state, the Player objects, a Battle object (if one is going on), and the saved turn information. Given this, we should be able to exactly calculate what moves are legal, whose turn it is, and what should be done next.