finished Marquise Action function, Eyrie's started
Finished most of the work needed to carry out any action by the Marquise. It's crazy how many different little things need to be tracked throughout a single turn! Also started work on the corresponding methods for the Eyrie.

It's still a little unclear where changing the phase_steps property should be done (action resolver or turn advancer methods), but it should just work in the end.

Will hopefully move on to the grand challenge: the turn advancement function.