finished main game / started debugging and tests
Finished adding in Field Hospitals and a loop that will automatically run the game with random actions taken to help me debug.

I still need to check for the game to be over and still have quite a ways with generating an observation, but I am really proud of this progress!